修改：可以看和自己的差值

+++ src/main/java/com/now/nowbot/dao/BeatmapDao.kt
@@ -446,7 +446,7 @@ class BeatmapDao(
 
         fun fromBeatmapsetModel(mapSet: Beatmapset): BeatmapsetLite {
             val s = BeatmapsetLite()
-            s.id = Math.toIntExact(mapSet.beatmapsetID)
+            s.id = mapSet.beatmapsetID.toInt()
             s.card = mapSet.covers.card2x
             s.cover = mapSet.covers.cover2x
             s.list = mapSet.covers.list2x
